Criação de sistema de bolsa de valores

+++ src/main/java/com/kuuhaku/controller/postgresql/CardMarketDAO.java
@@ -116,9 +116,9 @@ public class CardMarketDAO {
 		Calendar cal = Calendar.getInstance();
 		cal.add(Calendar.MONTH, -1);
 
-		double average;
+		int average;
 		try {
-			average = (int) em.createQuery("""
+			average = (int) (double) em.createQuery("""
 					SELECT AVG(cm.price)
 					FROM CardMarket cm
 					WHERE cm.card.card = :card
@@ -155,9 +155,9 @@ public class CardMarketDAO {
 	public static double getStockValue(Card c) {
 		EntityManager em = Manager.getEntityManager();
 
-		double average;
+		int average;
 		try {
-			average = (int) em.createQuery("""
+			average = (int) (double) em.createQuery("""
 					SELECT AVG(cm.price)
 					FROM CardMarket cm
 					WHERE cm.card.card = :card
